Initial test OK
This program tests the functionality of the module. It spins a motor at full speed. The output voltage of the L298N motor controller was measured to be almost 11V, which hopefully should provide enough power to run the pump at a satisfactory speed.  The next step is to plan the general structure of the program, and to make my own motor controller library.